Update get delete put endpoints

+++ src/products/products.controller.ts
@@ -3,6 +3,7 @@ import { ProductsService } from './products.service';
 import { CreateProductDTO } from './dtos/create-product.dto';
 import { ParseUUIDPipe } from '@nestjs/common';
 import { UpdateProductDTO } from './dtos/update-product.dto';
+import { NotFoundException } from '@nestjs/common';
 
 @Controller('products')
 export class ProductsController {
@@ -15,11 +16,15 @@ getAll(): any {
 
 @Get('/:id')
 getById(@Param('id', new ParseUUIDPipe()) id: string) {
-    return this.productsService.getById(id);
+  const prod = this.productsService.getById(id);
+  if (!prod) throw new NotFoundException('Product not found');
+  return prod;
 }
 
 @Delete('/:id')
 deleteById(@Param('id', new ParseUUIDPipe()) id: string) {
+  if (!this.productsService.getById(id))
+    throw new NotFoundException('Product not found');
   this.productsService.deleteById(id);
   return { success: true };
 }
@@ -30,12 +35,15 @@ create(@Body() productData: CreateProductDTO) {
 }
 
 @Put('/:id')
-update(
-  @Param('id', new ParseUUIDPipe()) id: string,
-  @Body() productData: UpdateProductDTO,
-) {
-  this.productsService.updateById(id, productData);
-  return { success: true };
-}
+  update(
+    @Param('id', new ParseUUIDPipe()) id: string,
+    @Body() productData: UpdateProductDTO,
+  ) {
+    if (!this.productsService.getById(id))
+      throw new NotFoundException('Product not found');
+
+    this.productsService.updateById(id, productData);
+    return { success: true };
+  }
 
 }
